# server/common/swagger/Api.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: myapp
  description: My cool Mongoose, Typescript, Express app
basePath: /api
...
consumes:
  - application/json
produces:
  - application/json
definitions: # ini buat definisikan model/schema data
...
  UserBody:
    type: object
    title: User
    required:
      - nip
      - name 
      - email 
      - password 
      - role 
      - jurusan
      - prodi
    properties:
      nip:
        type: string
        example: "0912831827577188"
      name:
        type: string
        example: "test"
      email:
        type: string
        example: test@gmail.com
      password:
        type: string
        example: "1234"
      role:
        type: string
        example: "admin"
      jurusan:
        type: string
        example: "JMTI"
      prodi:
        type: string
        example: "Informatika"
  MatkulBody:
    type: object
    title: Matkul
    required:
      - kd_matkul 
      - name 
      - sks 
      - pembagian
      - jenis 
      - prasyarat 
      - semester
      - praktikum
    properties:
      kd_matkul:
        type: string
        example: stress
      name:
        type: string
        example: stress
      sks:
        type: string
        example: stress
      pembagian:
        type: array
        items:
          type: number
        example: [1,2]
      jenis:
        type: string
        example: stress
      prasyarat:
        type: string
        example: stress
      semester:
        type: string
        example: stress
      praktikum:
        type: string
        example: stress
  JurusanBody:
    type: object
    title: Jurusan
    required:
      - kd_jurusan
      - name
    properties:
      kd_jurusan:
        type: string
        example: JAA
      name:
        type: string
        example: Jurusan Apa Apa
      prodi:
        type: array
        items:
          type: string
        example: []
  ProdiBody:
    type: object
    title: Prodi
    required:
      - kd_jurusan
      - prodi
    properties:
      kd_jurusan:
        type: string
        example: JAA
      prodi:
        type: array
        items:
          type: object
        example: ["prodi","prodi2"]

# url yang di akses user jadi http://localhost:3000/$basePath/paths
paths: 
...
  /user/{id}:
    get:
      tags:
        - User
      description: Get an user
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the user to retrieve
          type: string
      responses:
        200:
          description: Return the user with the specified id
        404:
          description: User not found
    patch:
      tags:
        - User
      description: Update an user
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the user to update
          type: string
        - name: user
          in: body
          description: user attribtues to update
          required: true
          schema: 
            $ref: "#/definitions/UserBody"
      responses:
        200:
          description: Return updated user
        404:
          description: User not found
    delete:
      tags:
        - User
      description: Delete an user
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the user to delete
          type: string
      responses:
        204:
          description: Delete the user with the specified id
        404:
          description: User not found

# matkul crud
...
  /matkul/{id}:
    get:
      tags:
        - Matkul
      description: Get an matkul
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the matkul to retrieve
          type: string
      responses:
        200:
          description: Return the matkul with the specified id
        404:
          description: Matkul not found
    patch:
      tags:
        - Matkul
      description: Update an matkul
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the matkul to update
          type: string
        - name: matkul
          in: body
          description: matkul attribtues to update
          required: true
          schema: 
            $ref: "#/definitions/MatkulBody"
      responses:
        200:
          description: Return updated matkul
        404:
          description: Matkul not found
    delete:
      tags:
        - Matkul
      description: Delete an matkul
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the matkul to delete
          type: string
      responses:
        204:
          description: Delete the matkul with the specified id
        404:
          description: Matkul not found

# jurusan crud
...
  /jurusan/{id}:
    get:
      tags:
        - Jurusan
      description: Get an jurusan
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the jurusan to retrieve
          type: string
      responses:
        200:
          description: Return the jurusan with the specified id
        404:
          description: Jurusan not found
    patch:
      tags:
        - Jurusan
      description: Update an jurusan
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the jurusan to update
          type: string
        - name: jurusan
          in: body
          description: jurusan attribtues to update
          required: true
          schema: 
            $ref: "#/definitions/JurusanBody"
      responses:
        200:
          description: Return updated jurusan
        404:
          description: Jurusan not found
    delete:
      tags:
        - Jurusan
      description: Delete an jurusan
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the jurusan to delete
          type: string
      responses:
        204:
          description: Delete the jurusan with the specified id
        404:
          description: Jurusan not found

# prodi crud
...
  /prodi/{id}:
    get:
      tags:
        - Prodi
      description: Get an prodi
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the prodi to retrieve
          type: string
      responses:
        200:
          description: Return the prodi with the specified id
        404:
          description: Prodi not found
    patch:
      tags:
        - Prodi
      description: Update an prodi
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the prodi to update
          type: string
        - name: prodi
          in: body
          description: prodi attribtues to update
          required: true
          schema: 
            $ref: "#/definitions/ProdiBody"
      responses:
        200:
          description: Return updated prodi
        404:
          description: Prodi not found
    delete:
      tags:
        - Prodi
      description: Delete an prodi
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the prodi to delete
          type: string
      responses:
        204:
          description: Delete the prodi with the specified id
        404:
          description: Prodi not found